+++ src/main/java/core/mvc/tobe/AnnotationHandlerMapping.java
@@ -4,11 +4,7 @@ import com.google.common.collect.Maps;
 import core.annotation.web.Controller;
 import core.annotation.web.RequestMapping;
 import core.annotation.web.RequestMethod;
-import core.di.BeanScanner;
 import core.di.ApplicationContext;
-import core.di.ComponentBasePackageScanner;
-import core.di.ConfigurationBeanScanner;
-import core.di.factory.BeanFactory;
 import core.mvc.HandlerMapping;
 import core.mvc.tobe.support.ArgumentResolver;
 import core.mvc.tobe.support.HttpRequestArgumentResolver;
